Fixes UIColor problems on custom buttons.  Custom UIToolbar doesn't work with iOS 11.

+++ Pacer/Constants.swift
@@ -159,6 +159,6 @@ struct Increment {
 }
 
 struct Colors {
-    static let Tint: UIColor = UIColor(red: CGFloat(57.0 / 255.0), green: CGFloat(149.0 / 255.0), blue: CGFloat(148.0 / 255.0), alpha: CGFloat(1.0)) // #399594
-    static let DarkTint: UIColor = UIColor(red: 58 / 255.0, green: 51 / 255.0, blue: 53 / 255.0, alpha: 1.0) // #3A3335
+    static let Tint: UIColor = UIColor(red: CGFloat(57.0/255.0), green: CGFloat(149.0/255.0), blue: CGFloat(148.0/255.0), alpha: CGFloat(1.0)) // #399594
+    static let DarkTint: UIColor = UIColor(red: CGFloat(58.0/255.0), green: CGFloat(51.0/255.0), blue: CGFloat(53.0/255.0), alpha: CGFloat(1.0)) // #3A3335
 }
